fix(calculateDotColour.ts): update import path for HabitColorGradients type in calculateDotColour.ts to match the new file location
fix(index.ts): update import path for CustomThemeColors type in index.ts to match the new file location

fix(useCustomTheme.tsx): update import path for CustomThemeColors type in useCustomTheme.tsx to match the new file location
